requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
Django==3.2.15
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
Markdown==3.3.7
multidict==6.0.2
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
yarl==1.7.2
requests==2.27.1